--- lesson-11/src/App.js
@@ -1,23 +1,0 @@
-import './App.css';
-import { useState } from 'react'
-
-// useState 
-
-function App() {
-  const [name, setName] = useState("Ahror")
- 
-  
-  
-  const handleClick = () => {
-    setName('Doniyor')
-  }
-
-  return (
-    <div className="App">
-        <h1>My name is {name}</h1>
-        <button onClick={handleClick}>Change name</button>
-    </div>
-  );
-}
-
-export default App;
